+++ src/Eloquent/Builder.php
@@ -13,6 +13,7 @@ class Builder extends EloquentBuilder
 
     /**
      * The methods that should be returned from query builder.
+     *
      * @var array
      */
     protected $passthru = [
@@ -191,7 +192,8 @@ class Builder extends EloquentBuilder
      * TODO Remove if https://github.com/laravel/framework/commit/6484744326531829341e1ff886cc9b628b20d73e
      * wiil be reverted
      * Issue in laravel frawework https://github.com/laravel/framework/issues/27791.
-     * @param array $values
+     *
+     * @param  array  $values
      * @return array
      */
     protected function addUpdatedAtColumn(array $values)
@@ -215,5 +217,28 @@ class Builder extends EloquentBuilder
     public function getConnection()
     {
         return $this->query->getConnection();
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function ensureOrderForCursorPagination($shouldReverse = false)
+    {
+        if (empty($this->query->orders)) {
+            $this->enforceOrderBy();
+        }
+
+        if ($shouldReverse) {
+            $this->query->orders = collect($this->query->orders)->map(function ($direction) {
+                return $direction === 1 ? -1 : 1;
+            })->toArray();
+        }
+
+        return collect($this->query->orders)->map(function ($direction, $column) {
+            return [
+                'column' => $column,
+                'direction' => $direction === 1 ? 'asc' : 'desc',
+            ];
+        })->values();
     }
 }