+++ tests/QueryTest.php
@@ -383,6 +383,29 @@ class QueryTest extends TestCase
         $this->assertEquals(1, $results->currentPage());
     }
 
+    public function testCursorPaginate(): void
+    {
+        $results = User::cursorPaginate(2);
+        $this->assertEquals(2, $results->count());
+        $this->assertNotNull($results->first()->title);
+        $this->assertNotNull($results->nextCursor());
+        $this->assertTrue($results->onFirstPage());
+
+        $results = User::cursorPaginate(2, ['name', 'age']);
+        $this->assertEquals(2, $results->count());
+        $this->assertNull($results->first()->title);
+
+        $results = User::orderBy('age', 'desc')->cursorPaginate(2, ['name', 'age']);
+        $this->assertEquals(2, $results->count());
+        $this->assertEquals(37, $results->first()->age);
+        $this->assertNull($results->first()->title);
+
+        $results = User::whereNotNull('age')->orderBy('age', 'asc')->cursorPaginate(2, ['name', 'age']);
+        $this->assertEquals(2, $results->count());
+        $this->assertEquals(13, $results->first()->age);
+        $this->assertNull($results->first()->title);
+    }
+
     public function testUpdate(): void
     {
         $this->assertEquals(1, User::where(['name' => 'John Doe'])->update(['name' => 'Jim Morrison']));